# specs/001-erp-core/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: WellKorea ERP API
  description: |
    REST API for WellKorea Integrated Work System.
    Manages Projects (identified by JobCode), quotations, approvals, production tracking, deliveries, invoicing, and payments.

    **Key Concepts**:
    - **Project**: Core domain entity representing a customer work request
    - **JobCode**: Unique business identifier for each Project (format: WK2{year}-{sequence}-{date})
    - **Approval**: Separate domain handling approval workflows for quotations and other entities

    ## CQRS Pattern

    This API follows the **Command Query Responsibility Segregation (CQRS)** pattern:

    ### Command Endpoints (Write Operations)
    Command endpoints (POST for create/submit, PUT for update) return a minimal `CommandResult`:
    ```json
    {
      "id": 123,
      "message": "Quotation created successfully"
    }
    ```
    After a command succeeds, clients should fetch the updated entity via the corresponding query endpoint if they need fresh data.

    ### Query Endpoints (Read Operations)
    Query endpoints (GET) return full entity representations with all fields.

    ### Benefits
    - **Simplicity**: Commands don't need to assemble complex response objects
    - **Consistency**: Clients always get the latest state from query endpoints
    - **Scalability**: Commands and queries can be scaled independently
    - **Event-driven**: Commands can publish domain events for side effects (e.g., email notifications)
...
servers:
...
paths:
...
  /approvals:
    get:
      tags:
        - Approval
      summary: List approval requests (filtered by status, entity type)
      operationId: listApprovals
      security:
        - BearerAuth: [ ]
      parameters:
        - name: entity_type
          in: query
          schema:
            type: string
            enum: [ Quotation, PurchaseOrder ]
        - name: status
          in: query
          schema:
            type: string
            enum: [ Pending, Approved, Rejected ]
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of approval requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApprovalRequestResponse'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'


    post:
      tags:
        - Approval
      summary: Create approval request for an entity
      description: |
        Submit an entity (Quotation, PurchaseOrder, etc.) for approval.
        Creates an ApprovalRequest with status=Pending.
      operationId: createApprovalRequest
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - entity_type
                - entity_id
              properties:
                entity_type:
                  type: string
                  enum: [ Quotation, PurchaseOrder ]
                entity_id:
                  type: string
                  format: uuid
                comments:
                  type: string
                  description: Optional submission comments
      responses:
        '201':
          description: Approval request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalRequestResponse'
        '400':
          description: Entity cannot be submitted for approval (validation errors)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Entity not found
  /approvals/{id}:
    get:
      tags:
        - Approval
      summary: Get approval request details
      operationId: getApprovalRequest
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Approval request details with history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalRequestResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /approvals/{id}/approve:
    post:
      tags:
        - Approval
      summary: Approve request at current level (CQRS Command)
      description: |
        Approve the request at the current level. Only the expected approver at the
        current level can approve. After approval:
        - If more levels remain: current_level increments, next approver can act
        - If this was the last level: status changes to APPROVED

        Returns CommandResult with the approval request ID.
        Use GET /approvals/{id} to fetch the updated approval status.
      operationId: approveRequest
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                comments:
                  type: string
                  description: Optional approval comments
      responses:
        '200':
          description: Request approved at current level (CQRS - returns ID only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only the expected approver at current level can approve
        '404':
          $ref: '#/components/responses/NotFound'
  /approvals/{id}/reject:
    post:
      tags:
        - Approval
      summary: Reject request at current level (CQRS Command)
      description: |
        Reject the approval request at the current level. Only the expected approver
        at the current level can reject. Rejection stops the workflow immediately.

        Returns CommandResult with the approval request ID.
        Use GET /approvals/{id} to fetch the updated approval status.
      operationId: rejectRequest
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Mandatory rejection reason
                comments:
                  type: string
                  description: Optional additional comments
              required:
                - reason
      responses:
        '200':
          description: Request rejected (CQRS - returns ID only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
        '400':
          description: Rejection reason is mandatory
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only the expected approver at current level can reject
        '404':
          $ref: '#/components/responses/NotFound'

  # Approval Chain Configuration Endpoints (Admin)
  # ============================================================================
  /admin/approval-chains:
    get:
      tags:
        - Admin
      summary: List all approval chain templates
      description: Returns all approval chain configurations (one per entity type). Admin only.
      operationId: listApprovalChains
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: List of approval chain templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApprovalChainTemplateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only Admin can view approval chain configuration
  /admin/approval-chains/{id}:
    get:
      tags:
        - Admin
      summary: Get approval chain by ID
      description: Returns the approval chain configuration for a specific template. Admin only.
      operationId: getApprovalChain
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Approval chain template with levels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalChainTemplateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only Admin can view approval chain configuration
        '404':
          description: Chain template not found
  /admin/approval-chains/{id}/levels:
    put:
      tags:
        - Admin
      summary: Configure approval levels for chain template (CQRS Command)
      description: |
        Configure the approval chain levels for a specific template.
        Each level specifies a position name (팀장, 부서장, 사장) and the user assigned to that position.
        Admin only endpoint.

        Returns CommandResult with the chain template ID.
        Use GET /admin/approval-chains/{id} to fetch the updated configuration.
      operationId: updateApprovalChainLevels
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - levels
              properties:
                levels:
                  type: array
                  items:
                    type: object
                    required:
                      - level_order
                      - level_name
                      - approver_user_id
                    properties:
                      level_order:
                        type: integer
                        minimum: 1
                        description: Execution order (1 = first approver)
                      level_name:
                        type: string
                        description: Position title (팀장, 부서장, 사장)
                      approver_user_id:
                        type: integer
                        format: int64
                        description: User ID of the approver at this level
                      is_required:
                        type: boolean
                        default: true
                        description: Whether this level can be skipped
      responses:
        '200':
          description: Approval chain levels updated (CQRS - returns ID only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
        '400':
          description: Invalid level configuration
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only Admin can configure approval chains
        '404':
          description: Chain template not found

  # Work Progress Endpoints
  # ============================================================================
...
components:
  schemas:
...
    CommandResult:
      type: object
      description: |
        Result of a command operation (CQRS pattern).
        Returns only the entity ID - clients should fetch fresh data via query endpoints.
      properties:
        id:
          type: integer
          format: int64
          description: ID of the created/updated entity
        message:
          type: string
          description: Human-readable success message
      required:
        - id
        - message
      example:
        id: 123
        message: "Quotation created successfully"

    # ========================================================================
    # Project Schemas
    # ========================================================================
...
    ApprovalChainTemplateResponse:
      type: object
      description: Approval chain configuration for an entity type
      properties:
        id:
          type: string
          format: uuid
        entity_type:
          type: string
          enum: [ QUOTATION, PURCHASE_ORDER ]
        name:
          type: string
          description: Human-readable name (견적서 결재)
        description:
          type: string
        is_active:
          type: boolean
        levels:
          type: array
          items:
            $ref: '#/components/schemas/ApprovalChainLevelResponse'
    ApprovalChainLevelResponse:
      type: object
      description: Approval level within a chain (팀장, 부서장, 사장)
      properties:
        id:
          type: string
          format: uuid
        level_order:
          type: integer
          description: Execution order (1 = first approver)
        level_name:
          type: string
          description: Position title (팀장, 부서장, 사장)
        approver_user_id:
          type: string
          format: uuid
        approver_user_name:
          type: string
          description: Full name of the approver
        is_required:
          type: boolean
    ApprovalRequestResponse:
      type: object
      description: Approval request instance with multi-level status
      properties:
        id:
          type: string
          format: uuid
        entity_type:
          type: string
          enum: [ QUOTATION, PURCHASE_ORDER ]
        entity_id:
          type: string
          format: uuid
        entity_description:
          type: string
          description: Human-readable description (견적서 v3 - WK2K25-0001-0104)
        current_level:
          type: integer
          description: Which level is currently awaiting approval
        total_levels:
          type: integer
          description: Total number of levels in the chain
        status:
          type: string
          enum: [ Pending, Approved, Rejected ]
          description: PENDING until all levels approve
        submitted_by_id:
          type: string
          format: uuid
        submitted_by_name:
          type: string
        submitted_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          description: When final approval/rejection occurred
        level_decisions:
          type: array
          description: Decision status at each level
          items:
            $ref: '#/components/schemas/ApprovalLevelDecisionResponse'
        history:
          type: array
          items:
            $ref: '#/components/schemas/ApprovalHistoryResponse'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/ApprovalCommentResponse'
    ApprovalLevelDecisionResponse:
      type: object
      description: Decision at each level of the approval chain
      properties:
        id:
          type: string
          format: uuid
        level_order:
          type: integer
        level_name:
          type: string
          description: Position title at this level
        expected_approver_id:
          type: string
          format: uuid
        expected_approver_name:
          type: string
        decision:
          type: string
          enum: [ Pending, Approved, Rejected ]
        decided_by_id:
          type: string
          format: uuid
        decided_by_name:
          type: string
        decided_at:
          type: string
          format: date-time
        comments:
          type: string
    ApprovalHistoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        level_order:
          type: integer
          description: Which level this action was for (null for SUBMITTED)
        action:
          type: string
          enum: [ Submitted, Approved, Rejected ]
        actor_id:
          type: string
          format: uuid
        actor_name:
          type: string
        timestamp:
          type: string
          format: date-time
        comments:
          type: string
    ApprovalCommentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        commenter_id:
          type: string
          format: uuid
        commenter_name:
          type: string
        comment_text:
          type: string
        created_at:
          type: string
          format: date-time
        is_rejection_reason:
          type: boolean
          description: True if this comment is a mandatory rejection reason

    # Work Progress Schemas
    # ========================================================================
...
  responses:
...
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: [ ]